Minor clean up and reorganizing of directory structure
 - Rename package elasticsearch => app
 - Rename elasticsearch-cli.go => main.go
 - Change interactive prompt to elasticsearch>
 - Version now prints Elasticsearch CLI
