+++ app/version.go
@@ -1,4 +1,4 @@
-package elasticshell
+package app
 
 const Version = "0.1.0"
